Fix undefined tutorialDetails
Because we register a user for all tutorials under a course when we have them start that course, we don't have any way to account for tutorials that are added AFTER a user registers for a course(there's no UserTutorials entry for the given tutorial).

1. Create a course as a teacher.
2. Create a course as a tutorial under that course.
3. Register for that course hitting Start From Beginning as another user.
4. Create another tutorial under that course.

From here, two issues pop up.
5a. call api/Tutorials/UpdateUserTutorial/[:tutorialId] . See it fail.

5b. Call api/GetUserTutorialsDetails/[:courseId] . See it fail to include the new tutorial from step 4.